requirements.txt
nicegui
fastapi
python-dotenv
google-api-python-client
google-auth
qrcode
qrcode[pil]
pydantic[email]
PyJWT
fastapi
fastapi[standard]
aioapns
google-cloud-storage
google-auth
python-dotenv
sqlalchemy
asyncpg
pydantic[email]
httpx
aiosmtplib
aiofiles
jinja2
google-api-python-client
tzdata
